sandbox/dan2: various modifications and bug fixed to support CUDA training in Dan's recipe.  Modify CUDA device selection code (cleanup, and make error-status controllable).
git-svn-id: https://svn.code.sf.net/p/kaldi/code/sandbox/dan2@3105 5e6a8d80-dfce-4ca6-a32a-6e07a63d50c8

+++ src/nnet2bin/nnet-gradient.cc
@@ -74,8 +74,9 @@ int main(int argc, char *argv[]) {
     }
 
     KALDI_ASSERT(minibatch_size > 0);
-    
-    int64 num_examples = 0;
+
+
+    double num_examples = 0.0;
     SequentialNnetTrainingExampleReader example_reader(examples_rspecifier);
 
     AmNnet am_gradient(am_nnet);
